+++ src/test/extension.test.ts
@@ -10,7 +10,7 @@ suite('Extension Test Suite', () => {
 
 	test('Configuration', () => {
 		const configuration = vscode.workspace.getConfiguration();
-		assert.strictEqual(0.5, configuration.get<number>("markdown.copilot.decorations.inactiveContextOpacity"));
+		assert.strictEqual(0.5, configuration.get<number>("markdown.copilot.context.inactiveOpacity"));
 		assert.strictEqual("", configuration.get<string>("markdown.copilot.backend.apiKey"));
 		assert.strictEqual("gpt-4o", configuration.get<string>("markdown.copilot.options.model"));
 		assert.strictEqual(0.1, configuration.get<number>("markdown.copilot.options.temperature"));